Refactor how CSV records are read
By moving the knowledge of how CSV record extraction works out of the
function that reads records, I get a function that should be easier to
unit test, and I believe the code is also simpler.

Also, some unused test_helper code commented out, for removal later.
